closes #2
traced the problem down to a change in the `crypto:hmac()` function's `$algorithm` parameter. as of expath-crypto-exist-lib-0.2.6 (in use on hsg production and svn trunk) we used `SunJCE` as the value for this parameter. but I noticed in https://github.com/claudius108/expath-exist/blob/master/crypto/src/test/java/org/expath/exist/crypto/xquery/hmacStringWithSha1.xq, @claudius108 used `base64` as the value. i tried it, and with the addition of the `http` namespace declarations (which also were not necessary in our 2.2-era code), the s3 code began working once again.